Break up types.d
Centralizing these types was a mistake, moving them to relevant components.
