+++ browser/src/ConstraintTable/ConstraintTable.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Gene, Transcript } from '../types'
+import { Gene } from '../GenePage/GenePage'
 
 import Link from '../Link'
 
@@ -62,9 +62,8 @@ const ConstraintTable = ({ datasetId, geneOrTranscript }: Props) => {
     return <p>Constraint not yet available for gnomAD v3.</p>
   }
 
-  const { transcriptId, transcriptVersion, transcriptDescription } = transcriptDetails(
-    geneOrTranscript
-  )
+  const { transcriptId, transcriptVersion, transcriptDescription } =
+    transcriptDetails(geneOrTranscript)
 
   const gnomadConstraint = geneOrTranscript.gnomad_constraint
   const exacConstraint = geneOrTranscript.exac_constraint
